refactor: reduce diff
